Final major edits and updates
Fixed error handlers. Added a color line to css to change the color of the project titles which were set to standard google purple once clicked. Updated error.pug. Updated page not found error (will change the picture to something user friendly). Added a special code in project.pug which allowed me to iterate through and post the two images to my projects, but also the 5 images for one of my projects. (Created variable set at zero, while 0 is less than the length of the project images then it will post the images attached to that project and will not post any space holders - it just posts what I have listed for that project regardless if it's two or five). This code was provided to me courtesy of Student Support Specialist, Brian - to replace my code that only populated the required two pics per project.